feat(OffboardingDashboard): cast days remaining to integer for accurate calculations
fix(migration): skip migration if deprecated leave policies are missing

feat(seeder): add automated backup jobs to CronJobSeeder

docs: add user manual by role for HRIS system

+++ app/Http/Controllers/HR/Offboarding/OffboardingDashboardController.php
@@ -374,7 +374,7 @@ class OffboardingDashboardController extends Controller
             ? round(($case->clearanceItems->where('status', 'approved')->count() / $case->clearanceItems->count()) * 100)
             : 0;
 
-        $daysRemaining = now()->diffInDays($case->last_working_day, false);
+        $daysRemaining = (int) now()->diffInDays($case->last_working_day, false);
         $isOverdue = $daysRemaining < 0;
 
         return [
